refactor(types/schemas): rename schema/ to schemas/, add index
